First steps to support HiFiBerryOS
The ongoing plan is to adopt HiFiBerryOS as the underlying platform for
Beocreate 2. This update includes changes to support this, but the work
is not complete:
- The DSP module from Beocreate Essentials is no longer referenced in
the main server, it should instead be included in each extension that
needs it.
- Settings should now be stored at /etc/beocreate/
- A configuration file (default in "defaults") allows defining the
sound card type (this is managed by HiFiBerryOS) and the port to use
for the HTTP/WS server.
- The UI and Beocreate Connect have been updated to support other ports
than 80.