Better juggling of active sources (?)
If a source starts playing it is now "reactivated", regardless of if it
was stopped or paused before, permitting display of correct metadata.
Additionally, start the server with "vv" argument to log an overview of
sources when the active source changes.

Introduces a new verbosity level 3 (vvv), to which the API endpoint
request logging has been moved.