When a DSP program is auto-installed, call HiFiBerry system reconfiguration if needed
When the system boots the first time on the new master board when
setting up daisy-chaining, the correct DSP program may not be
installed. This means the system will configure itself for software
volume control. Beocreate 2 will install the correct program, but the
HiFiBerry reconfiguration script needs to run to update the available
DSP mixer to all sources.